New proposed docs for VSC structured
I read through the old VSC docs and a lot of the stuff has been marked as outdated (thread with caution) so I didnt add any of that. This version definitely lacks it and will need a dev to update each page. But its set up so its easy to expand on.